refactor(upload): constantes de dépôt et conversion IGNF→EPSG mutualisées
