+++ assets/utils/upload.ts
@@ -1,3 +1,16 @@
+import ignfProjections from "@/data/ignf_projections.json";
+
+/** contraintes Entrepôt sur le fichier de données déposé */
+export const DATASET_MAX_FILE_SIZE = 2_000_000_000; // 2 Go
+export const DATASET_FILE_EXTENSIONS = ["gpkg", "zip", "geojson", "csv", "sql"];
+
+/** convertit un code de projection IGNF en son équivalent EPSG quand il est connu ; undefined si aucune projection exploitable */
+export const mapIgnfToEpsg = (sridRaw: unknown): string | undefined => {
+    const srid = typeof sridRaw === "string" && sridRaw !== "" && sridRaw in ignfProjections ? ignfProjections[sridRaw] : sridRaw;
+
+    return typeof srid === "string" && srid.trim() !== "" ? srid : undefined;
+};
+
 /** parse le tag integration_progress d’une livraison (objet JSON étape → statut) */
 export const parseIntegrationProgress = (rawProgress: string | undefined): Record<string, string> | null => {
     if (!rawProgress) {
